Create pages folder and relocate applicantInformation there
