+++ src/js/edu-benefits/1995/config/form.js
@@ -15,7 +15,7 @@ import * as toursOfDuty from '../../definitions/toursOfDuty';
 import createContactInformationPage from '../../pages/contactInformation';
 import createOldSchoolPage from '../../pages/oldSchool';
 import createDirectDepositChangePage from '../../pages/directDepositChange';
-import createApplicantInformationPage from '../../../common/schemaform/definitions/applicantInformation';
+import createApplicantInformationPage from '../../../common/schemaform/pages/applicantInformation';
 
 import { showSchoolAddress } from '../../utils/helpers';
 import { benefitsLabels } from '../../utils/labels';
