+++ test/common/schemaform/definitions/applicantInformation.unit.spec.jsx
@@ -5,7 +5,7 @@ import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
 import { DefinitionTester, submitForm } from '../../../util/schemaform-utils.jsx';
-import applicantInformation from '../../../../src/js/common/schemaform/definitions/applicantInformation.js';
+import applicantInformation from '../../../../src/js/common/schemaform/pages/applicantInformation';
 
 import fullSchema1990e from 'vets-json-schema/dist/22-1990E-schema.json';
